# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Subscription Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: ok
  /api/v1/subscriptions/:
    post:
      summary: Create subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionPayload'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
    get:
      summary: List subscriptions
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{id}:
    get:
      summary: Get subscription by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found
    put:
      summary: Update subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionPayload'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
  /api/v1/subscriptions/total:
    get:
      summary: Calculate total subscription cost for period
      parameters:
        - in: query
          name: period_start
          required: true
          schema:
            type: string
            example: 01-2025
        - in: query
          name: period_end
          required: true
          schema:
            type: string
            example: 03-2025
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          schema:
            type: string
      responses:
        '200':
          description: Calculated
          content:
            application/json:
              schema:
                type: object
                properties:
                  period_start:
                    type: string
                  period_end:
                    type: string
                  total_price:
                    type: integer
  /api/v1/analytics/simulate:
    post:
      summary: Simulate restaurant demand analytics daily report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsSimulationRequest'
      responses:
        '200':
          description: Simulated report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsDailyReport'
components:
  schemas:
    AnalyticsSimulationRequest:
      type: object
      required: [sales_24_months, context]
      properties:
        sales_24_months:
          type: array
          items:
            $ref: '#/components/schemas/SaleRecord'
        ingredient_requirements:
          type: array
          items:
            $ref: '#/components/schemas/IngredientRequirement'
        stock:
          type: array
          items:
            $ref: '#/components/schemas/StockItem'
        context:
          $ref: '#/components/schemas/ExternalContext'
        buffer_percent:
          type: number
          format: float
        historical_plan_fact_points:
          type: array
          items:
            $ref: '#/components/schemas/PlanFactPoint'
    SaleRecord:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        dish_id:
          type: string
        dish_name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: float
        temperature:
          type: number
          format: float
        is_holiday:
          type: boolean
        event_factor:
          type: number
          format: float
    IngredientRequirement:
      type: object
      properties:
        dish_id:
          type: string
        ingredient_id:
          type: string
        ingredient:
          type: string
        quantity_per_dish:
          type: number
          format: float
        uom:
          type: string
    StockItem:
      type: object
      properties:
        ingredient_id:
          type: string
        ingredient:
          type: string
        quantity:
          type: number
          format: float
        uom:
          type: string
    ExternalContext:
      type: object
      properties:
        date:
          type: string
          format: date-time
        forecast_temperature:
          type: number
          format: float
        is_holiday:
          type: boolean
        event_name:
          type: string
        event_multiplier:
          type: number
          format: float
    PlanFactPoint:
      type: object
      properties:
        label:
          type: string
        plan:
          type: number
          format: float
        fact:
          type: number
          format: float
    AnalyticsDailyReport:
      type: object
      properties:
        date:
          type: string
        total_revenue:
          type: number
          format: float
        total_orders:
          type: integer
        by_dish:
          type: array
          items:
            $ref: '#/components/schemas/ForecastItem'
        by_category:
          type: array
          items:
            $ref: '#/components/schemas/CategoryForecast'
        hourly_load:
          type: array
          items:
            $ref: '#/components/schemas/HourForecast'
        purchase_plan:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseItem'
        plan_fact:
          $ref: '#/components/schemas/PlanFactMetrics'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/BusinessRecommendation'
    ForecastItem:
      type: object
      properties:
        dish_id:
          type: string
        dish_name:
          type: string
        category:
          type: string
        predicted_quantity:
          type: integer
        predicted_revenue:
          type: number
          format: float
    CategoryForecast:
      type: object
      properties:
        category:
          type: string
        predicted_quantity:
          type: integer
        predicted_revenue:
          type: number
          format: float
    HourForecast:
      type: object
      properties:
        hour:
          type: integer
        predicted_orders:
          type: integer
    PurchaseItem:
      type: object
      properties:
        ingredient_id:
          type: string
        ingredient:
          type: string
        need:
          type: number
          format: float
        stock:
          type: number
          format: float
        purchase:
          type: number
          format: float
        uom:
          type: string
    PlanFactMetrics:
      type: object
      properties:
        mape:
          type: number
          format: float
        wape:
          type: number
          format: float
        bias:
          type: number
          format: float
        rmse:
          type: number
          format: float
    BusinessRecommendation:
      type: object
      properties:
        title:
          type: string
        reason:
          type: string
        action:
          type: string
    SubscriptionPayload:
      type: object
      required: [service_name, price, user_id, start_date]
      properties:
        service_name:
          type: string
          example: Yandex Plus
        price:
          type: integer
          example: 400
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
          example: 07-2025
        end_date:
          type: string
          nullable: true
          example: 12-2025
    Subscription:
      allOf:
        - $ref: '#/components/schemas/SubscriptionPayload'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
            updated_at:
              type: string